uibuild

+++ typeDefs/types/chartDataType.js
@@ -10,7 +10,7 @@ type ChartData {
   time: [Date!]!
   temperature_C: [Float!]!
   ec_mS_cm: [Float]
-  light_lux: [Int]
+  light_lux: [Float]
   CO2_ppm: [Int]
   soil_moisture: [Int]
   humidity: [Int]
